Fix Bug: Add an alias 'decoration_id' to museum_object column 'decoration_style_id'.
As the name of Decoration changed for the user (decoration style), the naming convention inside MuseumObject class is not consistent with actual DB schema. Fulltext search did throw an error when it encountered a match for a term of class Decoration, as it afterward tried to join tables with museum_objects.decoration_id, which does not exist.