Remove obsolete app-loader application
As the application was used locally on a single laptop in phase 1, we created a qt based loader to manage updates and provide feedback to the user about the application. The application is now used as intended, as a web application within a server client context, this application has become obsolete.