Fix: App's absolute path has a lesser depth now
When retrieving the backup, we want to strip necessary sub directories as borg saves the absolute path of all files but uses it as relative paths. The path of the data was `home/dojam/jamappv2/app/` and `home/dojam/jamappv2/db/` but is now `root/dojam` and `root/db` and therefore we have to strip only the first two component now.