Fix #69: Enforce downcase capitalization rules
We want downcase at all places like form texts or titles. This is achieved by overriding both the `humanize` as well as `titleize` method of Rails' String extensions.
For `humanize` we just call the original with flipped default for the corresponding named parameter.
For `titleize` we just downcase the result. We also introduce a new named parameter to provide backward compatibility as the original titleize method is used internally at some places. Note that this is discouraged, titleize is supposed to be a pure visual decorator.
As we changed the `titleize` default behavior all places within the application where the method was called were changed to reproduce the old behavior of the method.