Use a docker volume for fast bootsnap startup
Bootsnap provides a faster startup of ruby/rails by caching, which for us is mostly important when testing.
However, the cache consists of many files and without that new volume all that read/write operations needed to be synced by docker to the local filesystem which made the startup extremely slow.