Fix Bug: Create all undetermined paths when creating main account
Having those paths is crucial for the app to work correctly. Up to now only the path of depth 4 (/undet/undet/undet/undet) was created on the fly if neccessary, but without creating it's parent paths. This caused issues when path was assigned as secondary as it's parent did not exist.